+++ src/main/kotlin/lotto/domain/LottoNumbers.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class LottoNumbers(val numbers: List<LottoNumber>) {
