+++ src/test/kotlin/lotto/LottoNumberTest.kt
@@ -1,5 +1,6 @@
 package lotto
 
+import lotto.domain.LottoNumber
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
