+++ src/test/kotlin/lotto/LottoNumbersTest.kt
@@ -1,5 +1,7 @@
 package lotto
 
+import lotto.domain.LottoNumber
+import lotto.domain.LottoNumbers
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
